requirements.txt
fastapi
uvicorn
pyyaml
ruamel.yaml